feature: execute command

+++ src/helpers/PdfHelper.zig
@@ -267,7 +267,7 @@ pub fn resetZoomAndScroll(self: *Self) void {
     self.x_offset = 0;
 }
 
-pub fn goToPage(self: *Self, pageNum: i32) void {
+pub fn goToPage(self: *Self, pageNum: u16) void {
     if (pageNum >= 0 and pageNum <= self.total_pages) {
         self.current_page_number = @as(u16, @intCast(pageNum));
     }
